refactor: rename helpers to stack
