+++ packages/relottie-parse/src/index.ts
@@ -5,4 +5,4 @@
 export { default, type Options } from './unified-relottie-parse.js';
 export type { ParseFileData } from './parse.js';
 export { DEFAULT_OPTIONS } from './options.js';
-export { Stack } from './helpers.js';
+export { Stack } from './stack.js';
