+++ packages/relottie-parse/src/stack.ts
